feat(ts): categorize commands
