+++ src/commands/game/higherlower.ts
@@ -8,10 +8,10 @@ import {
   MessageActionRowComponentBuilder,
 } from "discord.js";
 import { v4 as uuidv4 } from "uuid";
-import HOR from "../util/Classes/generateHOR";
-import { ChatInputCommand } from "../models";
-import { HigherlowerModel } from "../util/Models/higherlowerModel";
-import { getHigherLower } from "../util/Functions/jsonImport";
+import HOR from "../../util/Classes/generateHOR";
+import { ChatInputCommand } from "../../models";
+import { HigherlowerModel } from "../../util/Models/higherlowerModel";
+import { getHigherLower } from "../../util/Functions/jsonImport";
 
 const command: ChatInputCommand = {
   requireGuild: true,
