+++ src/commands/owner/reload.ts
@@ -1,7 +1,7 @@
 import { readdirSync } from "fs";
 import { captureException } from "@sentry/node";
 import { SlashCommandBuilder } from "discord.js";
-import { ChatInputCommand } from "../models";
+import { ChatInputCommand } from "../../models";
 const cat = readdirSync(`./src/commands/`).filter((d) => d.endsWith(".js"));
 
 const command: ChatInputCommand = {
