+++ src/commands/settings/language.ts
@@ -5,7 +5,7 @@ import {
   PermissionsBitField,
 } from "discord.js";
 import { captureException } from "@sentry/node";
-import { ChatInputCommand } from "../models";
+import { ChatInputCommand } from "../../models";
 
 const command: ChatInputCommand = {
   requireGuild: true,
